Model processor project and solution now compile fine with the new folder hierarchy and the new compile options (more restrictive than before).

+++ engine/src/code/revVideo/scene/model/staticModel.cpp
@@ -101,7 +101,7 @@ namespace rev { namespace video
 	}
 
 	//------------------------------------------------------------------------------------------------------------------
-	unsigned int CStaticModel::nTriangles() const
+	unsigned short CStaticModel::nTriangles() const
 	{
 		return mNTriangles;
 	}
